shibal/pyproject.toml
[tool.poetry]
name = "shibal"
version = "0.1.0"
description = ""
authors = ["Anton Fedotov <tosha.fedotov.2000@gmail.com>"]
license = "POEBAT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = {extras = ["mypy"], version = "^2.0.38"}
asyncpg = "^0.30.0"
fastapi = "^0.115.11"
redis = "^5.2.1"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
celery = {extras = ["redis"], version = "^5.4.0"}
pydantic-settings = "^2.8.1"
pydub = "^0.25.1"
python-telegram-bot = "^21.11.1"
moviepy = "^2.1.2"
httpx = "^0.28.1"
uuid = "^1.30"


[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ruff = "^0.9.9"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ['pydantic.mypy', 'sqlalchemy.ext.mypy.plugin']
